Modify input_util so that all plots can be saved with cmd line flags.
Previously only some plots were able to be saved. This meant the plots
were not accessible on systems that don't support the matplotlib
interactive viewer window. With this change there is now a cmdline
argument that is recognized by all the plotting funtions allowing an
image file to be saved.

The gap plot required re-organization so that rather than 2 figures, the
code produced 1 figure. This makes the cmd line parsing easier and the
cmd line options more intuitive with respect to the behavior.