Fix issue with build arguments being ignored for some images
Discovered that in some images, arugments passed to docker build were
not making their way into the proper substitutions in the Dockerfile.
This resulted in the default user being "root" in some cases as the USER
directives in the Dockerfile were being ignored (or run with a blank
UNAME).

Fixed to use the pattern where the default values/args are declared at
the top of the file, then in each stage the ARGs are declared again but
with no value. It then seems to work. The key is that ARGs are local to
a build stage and must be declared after each FROM statement.