FIX BUG with naming columns in the sample matrix.
previously the column names were not getting the PFT number appended. so
if you are working on cmax for several PFTs, you'd have duplicate column
names. This change adds the appropriate PFT suffix when the sample
matrix is created.

This fixes a downstream bug when the sample matrix is converted to a
dict for each row as it is passed into the core setup. With duplicate
column names, pandas.DataFrame.to_dict throws a warning but then simply
returns the last column name, and drops the other data.