Add code that can setup the SensitivityDriver outputs based on the targets...
It is intended that the targets are specified by the end user (for
example in a config file). After the user specifies the targets, this
setup_output(...) function can be called which will make a list of
output specifications so that the correct outputs are enabled based on
the user's chosen targets.