Adds initial files for defining test framework
This includes an initial example/pr test and
some utility methods, both for testing and
for general running.

The defined test is very preliminary and is likely to
change quite a bit as we test this approach.
The methods have limited error catching/handling
and need further development and polishing.