re-organize functions so that automatic cmd line docs can be generated.
The Sphinx extension that auto-documents the python command line
interfaces that are defined with argparse (sphinx-argparse) requires
that the script has a function that returns the parser object.

Here we re-factor the various command line parsing and entry scripts so
that we can provide such a function for all the files.

This basically amounts to putting the command line definition into into
its own function and then the cmdline_parse() function that previously
had the definition now simply has a call to the cmdline_define function.

Also adds the extension and a bunch of pages to the sphinx docs.