Start class hierarchy.
Now there is a base class so we can have other types of drivers.

Had to mess with the seedpath attribute...I had made this double
underscore for somereason, but this runs into issues with the name
mangling, when trying to access from the subclass. But if I make it
single under score it works...

Moves several methods the base class. More to follow as we figure
out where the overlap is between the Sensitivity and the MadsCA