Fix bug with SA: not setting site correctly in sample folders.
There was a significant bug with the Sensitivity driver class where
the site was not being set in the sample folders.

This commit changes the way the constructors work. Moves the
fromfilename class method from the MadsTEMDriver class to the Base
driver class. Makes the default constructor for all classes require a
configuration dictionary.

Fixes tests accordingly.