Sensitivity.py and SA_setup_and_run.py: change ctor and API
Significant change to ctor and API. Changing so that almost everything
is set thru the config dict passed to the ctor. This deprecates the
design_experiment(...) function. The idea is that the user will set
everything in the config.yaml file. The setup and run wrapper will
(eventually) provide some override options, but for not all options have
been removed. Would like to have a general pattern that the congig file
is used for all settings and then the user can optionally override some
things via command line interface.

This will require a number of keys to be added to the config file,
mostly to make up