Sensitivity: stub out functionality for equlibrium metrics.
This builds off the equlibrium_check.py code that is in the
mads_calibration directory. That code was not generalized, and required
adjustment for each run. This code is more generalized and might work
for a wide variety of cases.

This is a draft implementaiton, but it should run for a single variable
output. Has not been tested for non-PFT or multi variable situations.

Idea is that at the and of each run, in the "sample specific run folder"
(SSRF), more summary files are generated alongside the existing
results.csv. You will get one file for each enabled output variable. The
file will be a csv and will have several columns for each PFT, including
the following measurements of equlibrium:
 - the slope of a linear regression
 - the p value as returned from scipy.stats.lineregress
 - the cv (coefficient of variation?; calculated as was done in
   equlibrium_check.py)

It would be easy to add more metrics to this table (i.e. r, or std_err).

The csv file in your SSRF will have many columns and one row. The number
of coulumns will be (# of PFTs * # of metrics * # of compartments)

Then when all the runs are finished, these eq_{VARIABLE}_qaulity.csv
files are collected into a single file per variable with the same number
of columns and one row for each sample run. This final summary file will
be in the work_dir, alongside the other summary files.

Nothing has been added to the post hoc analysis, but the idea is that,
similar to the n top runs, you sort based on some metric and select only
the rows (sample runs) that pass the metric...

Additional changes with this commit:

 - add import
 - fix a few typos
 - improve error message
 - add more data to the info.txt file