fix bug with sample matrix column names.
Turns out that commit e9c6f864 was not a complete fix but simply made
the column names incorrect with the nmax/krb parameter set.

This fix should be more robust. It factors the column naming into a
function that can be re-used by different sampling methods and tries to
handle many possible NULL settings for pft number.