Change default behavior for plot saving, SA post hoc
With this commit, the common behavior is that each plot function takes a
saveprefix key word argument. It defaults to an empty string, so if the
user does not provide it, the plot is saved in the current working
directory with a default file name that reflects the function that made
the plot.
If the user provides a prefix, then the default name will be added to
the prefix.

Previously the user encountered errors if they did not have an existing
plots/ directory.