MadsTEMDriver.py: add another directory level to the work_dir...
...previously, the ``self.work_dir`` was used as the top level of the
run. Not we add a ``run`` directory, so that the setup is similar to the
Sensitivity Analysys (where all the runs take place in "sample specific
run folders"). Here we don't have a "sample specific run folder" but
putting the run itself in a dedicated folder means that summary level
plots and tables can be kept at the self.work_dir level and perhaps
persisted even if the ``self.work_dir/run/`` folder is cleaned out.