SA, CA, fix bug with calib_mode
The demo files used upper case "ALL", while the code expected 'All' in
the "GPPAllIgnoringNitrogen", which was confusing.

Here we standardize the capitalization in the demo files AND make the
code more flexible so that the user can pass any form of capitilization
and the string is cast to lowercase before being compared to see if the
modules should be turned on/off.

Also adds new test file.