fix plotting bug with AC-MADS-TEM.jl...
... the plotting has been failing beause the parameter opt_params was
not defined. This was an oversight in an attempt to accomodate both the
reading of optimum parameters the returned object of a successful
calibration or the .iterationresults files, which can be accessed after
the calibration code has exited.

Here we don't make an effort at reading the .iterationresults files, or
letting the use by-pass the time consuming calibraiton step and
opt_params is set to the values from the object returned from the
mads.calibrate(...) function call.

Also added commas to the println statements for easier copying directly
into the arguments of the Python util.metrics.plot_optimization_fit(...)
function.

And added some notes about accessing this data from a python
script.