Introduces an artificial 'time' entry for soilenv
The plotting calls need a list of integers, not
datetimes. The problem only arose for the code
sections that used water table, as that is a monthly
variable. All other sections simply used 'year'.
This result is inelegant, but does produce the results
needed. It should be replaced by something more
robust when possible.