Make additions to dvmdostem-dev docker image for using gcsfuse.
This installs the gcsfuse package(s) in the image, and changes
the compose file to add some settings that allow the gcsfuse to work.

Then user can mount with command like this:

gcsfuse --implicit-dirs <BUCKET-NAME> <PATH-TO-MOUNT>

Depending on the environment you might need to provide some credentials.
From a docker container on a GCP instance it seems to work. From local
machine it might require --key-file <path-to-key.json> flag.