Remove git LFS hashes.
Git LFS usage quotas are too low and the prices too expensive for
this project. Removing hashes (the hashes are the small files
that are actually tracked) from the project so that users are not
getting errors when cloning and checking out. Users may still get
errors if they try to checkout a point in the history tree with
the hash files tracked. This is surmountable by setting an
environment variable telling git-lfs not to run the smudge
filter, e.g.:

    $ GIT_LFS_SKIP_SMUDGE=1 git checkout <SHA>