Modify/improve MadsTEMDriver.observed_vec function, fix bug.
Adds the capability for the observed_vec function to return either
a labeled array or flat list of target (observation) values. Turns out
that the labeled array was helpful for testing/cross checking the order
of the flat list.

Adds tests that go with the new feature.

Fixes bug with indentation of return statement in observed_vec function.

Change AC-MADS-TEM.jl so that it explicitly requests a flat list as the
remainder of MADS seems to expect.