Change paths to use demo data for the demo config and testing data for the tests.
This is sort of confusing - in the process of re-arranging the way that
testing data is stored and handled (see pull request and branch removing
GitLFS). But the ideas is that the demo data should come with the repo
and the testing data will require an extra step (details yet to be
determined) for the user to acquire.