Fix issue with tests:
One of the tests was failing on a whitespace/formatting difference from
printing out the head/tail of a DataFrame. This is a weakness of the
doctest arrangement, but here we introduce the pytest library so we can
use the pytest.approx(..) function. While this makes the doctest less
enjoyable to read, it should be much less brittle.